OMP done

+++ CPP-TFS/FlowAlgorithm.hpp
@@ -7,7 +7,7 @@
 
 double calculate_path_weight(const Graph& g, const std::vector<int>& path);
 
-void flowDistributionAlgorithm(Graph& g, vector<Commodity>& commodities, double num_of_iter);
+void flowDistributionAlgorithm(Graph& g, vector<Commodity>& commodities, int num_of_iter);
 
 vector<boost::graph_traits<Graph>::edge_descriptor> get_edges_with_flow(Graph& g);
